Rename module i_gui_context -> i_event_loop
